landing-page-api/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
asyncpg==0.29.0
pydantic[email]==2.5.3
python-multipart==0.0.6
langchain==0.1.0
langchain-openai==0.0.2
python-dotenv==1.0.0
